# backend/src/routes/auth.js
const express = require('express');
const bcrypt = require('bcryptjs');
const jwt = require('jsonwebtoken');
const { getDb } = require('../database');
const { authenticate } = require('../middleware/auth');
const { validateEmail, validatePassword, validateRequired } = require('../utils/validators');
const { errorResponse, successResponse } = require('../utils/responses');

const router = express.Router();

function generateAccessToken(user) {
  return jwt.sign(
    { id: user.id, email: user.email, role: user.role },
    process.env.JWT_SECRET,
    { expiresIn: process.env.JWT_ACCESS_EXPIRY || '15m' }
  );
}

function generateRefreshToken(user) {
  return jwt.sign(
    { id: user.id, email: user.email, role: user.role },
    process.env.JWT_REFRESH_SECRET,
    { expiresIn: process.env.JWT_REFRESH_EXPIRY || '7d' }
  );
}

/**
 * POST /api/auth/register
 */
router.post('/register', (req, res) => {
  const { name, email, password } = req.body;

  // Validate required fields
  const requiredErrors = validateRequired(['name', 'email', 'password'], req.body);
  if (requiredErrors.length > 0) {
    return errorResponse(res, 422, 'Validation failed', 'VALIDATION_ERROR', requiredErrors);
  }

  // Validate name
  if (typeof name !== 'string' || name.trim().length < 2) {
    return errorResponse(res, 422, 'Name must be at least 2 characters', 'VALIDATION_ERROR');
  }

  // Validate email
  const emailError = validateEmail(email);
  if (emailError) {
    return errorResponse(res, 422, emailError, 'VALIDATION_ERROR');
  }

  // Validate password
  const passwordError = validatePassword(password);
  if (passwordError) {
    return errorResponse(res, 422, passwordError, 'VALIDATION_ERROR');
  }

  const db = getDb();

  // Check duplicate email
  const existing = db.prepare('SELECT id FROM users WHERE email = ?').get(email.trim().toLowerCase());
  if (existing) {
    return errorResponse(res, 409, 'Email already registered', 'DUPLICATE_EMAIL');
  }

  // Hash password and create user
  const hashedPassword = bcrypt.hashSync(password, 10);
  const result = db.prepare(
    'INSERT INTO users (name, email, password, role) VALUES (?, ?, ?, ?)'
  ).run(name.trim(), email.trim().toLowerCase(), hashedPassword, 'user');

  const user = db.prepare('SELECT id, name, email, role, created_at FROM users WHERE id = ?').get(result.lastInsertRowid);

  const accessToken = generateAccessToken(user);
  const refreshToken = generateRefreshToken(user);

  return successResponse(res, {
    message: 'Registration successful',
    user: { id: user.id, name: user.name, email: user.email, role: user.role },
    accessToken,
    refreshToken,
    tokenType: 'Bearer',
    expiresIn: process.env.JWT_ACCESS_EXPIRY || '15m',
  }, 201);
});

/**
 * POST /api/auth/login
 */
router.post('/login', (req, res) => {
  const { email, password } = req.body;

  // Validate required fields
  const requiredErrors = validateRequired(['email', 'password'], req.body);
  if (requiredErrors.length > 0) {
    return errorResponse(res, 422, 'Validation failed', 'VALIDATION_ERROR', requiredErrors);
  }

  // Validate email format
  const emailError = validateEmail(email);
  if (emailError) {
    return errorResponse(res, 422, emailError, 'VALIDATION_ERROR');
  }

  const db = getDb();
  const user = db.prepare('SELECT * FROM users WHERE email = ?').get(email.trim().toLowerCase());

  if (!user) {
    return errorResponse(res, 401, 'Invalid email or password', 'INVALID_CREDENTIALS');
  }

  const validPassword = bcrypt.compareSync(password, user.password);
  if (!validPassword) {
    return errorResponse(res, 401, 'Invalid email or password', 'INVALID_CREDENTIALS');
  }

  const accessToken = generateAccessToken(user);
  const refreshToken = generateRefreshToken(user);

  return successResponse(res, {
    message: 'Login successful',
    user: { id: user.id, name: user.name, email: user.email, role: user.role },
    accessToken,
    refreshToken,
    tokenType: 'Bearer',
    expiresIn: process.env.JWT_ACCESS_EXPIRY || '15m',
  });
});

/**
 * POST /api/auth/refresh
 */
router.post('/refresh', (req, res) => {
  const { refreshToken } = req.body;

  if (!refreshToken) {
    return errorResponse(res, 422, 'Refresh token is required', 'VALIDATION_ERROR');
  }

  // Check if refresh token is blacklisted
  const db = getDb();
  const blacklisted = db.prepare('SELECT id FROM blacklisted_tokens WHERE token = ?').get(refreshToken);
  if (blacklisted) {
    return errorResponse(res, 401, 'Refresh token has been revoked', 'TOKEN_REVOKED');
  }

  try {
    const decoded = jwt.verify(refreshToken, process.env.JWT_REFRESH_SECRET);

    // Verify user still exists
    const user = db.prepare('SELECT id, name, email, role FROM users WHERE id = ?').get(decoded.id);
    if (!user) {
      return errorResponse(res, 401, 'User not found', 'USER_NOT_FOUND');
    }

    const newAccessToken = generateAccessToken(user);
    const newRefreshToken = generateRefreshToken(user);

    // Blacklist the old refresh token
    const expDate = new Date(decoded.exp * 1000).toISOString();
    db.prepare('INSERT OR IGNORE INTO blacklisted_tokens (token, expires_at) VALUES (?, ?)').run(refreshToken, expDate);

    return successResponse(res, {
      accessToken: newAccessToken,
      refreshToken: newRefreshToken,
      tokenType: 'Bearer',
      expiresIn: process.env.JWT_ACCESS_EXPIRY || '15m',
    });
  } catch (err) {
    if (err.name === 'TokenExpiredError') {
      return errorResponse(res, 401, 'Refresh token has expired', 'TOKEN_EXPIRED');
    }
    return errorResponse(res, 401, 'Invalid refresh token', 'INVALID_TOKEN');
  }
});

/**
 * POST /api/auth/logout (requires token)
 */
router.post('/logout', authenticate, (req, res) => {
  const db = getDb();

  // Blacklist the current access token
  try {
    const decoded = jwt.decode(req.token);
    const expDate = new Date(decoded.exp * 1000).toISOString();
    db.prepare('INSERT OR IGNORE INTO blacklisted_tokens (token, expires_at) VALUES (?, ?)').run(req.token, expDate);
  } catch (e) {
    // Token already invalid, that's fine
  }

  // Also blacklist refresh token if provided
  const { refreshToken } = req.body;
  if (refreshToken) {
    try {
      const decoded = jwt.decode(refreshToken);
      if (decoded && decoded.exp) {
        const expDate = new Date(decoded.exp * 1000).toISOString();
        db.prepare('INSERT OR IGNORE INTO blacklisted_tokens (token, expires_at) VALUES (?, ?)').run(refreshToken, expDate);
      }
    } catch (e) {
      // ignore
    }
  }

  return successResponse(res, { message: 'Logout successful' });
});

/**
 * GET /api/auth/me (requires token)
 */
router.get('/me', authenticate, (req, res) => {
  const db = getDb();
  const user = db.prepare('SELECT id, name, email, role, created_at, updated_at FROM users WHERE id = ?').get(req.user.id);

  if (!user) {
    return errorResponse(res, 404, 'User not found', 'NOT_FOUND');
  }

  return successResponse(res, { user });
});

module.exports = router;